backend/requirements.txt
# Web Framework
Flask>=3.0.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0

# Environment Configuration
python-dotenv>=1.0.0,<2.0.0

# Stock Data & Processing
yfinance>=0.2.36
pandas>=2.0.0,<3.0.0

# Natural Language Processing (NLP)
spacy>=3.7.0,<4.0.0